backend/requirements.txt
appnope==0.1.3
asttokens==2.2.1
autopep8==2.0.0
awscli==1.27.24
backcall==0.2.0
bcrypt==4.0.1
blinker==1.5
boto3==1.26.24
botocore==1.29.24
click==8.1.3
colorama==0.4.4
decorator==5.1.1
dnspython==2.2.1
docutils==0.16
email-validator==1.3.0
executing==1.2.0
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-DebugToolbar==0.13.1
Flask-SQLAlchemy==3.0.2
Flask-WTF==1.0.1
greenlet==2.0.1
idna==3.4
ipython==8.7.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.33
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.4.8
pycodestyle==2.10.0
Pygments==2.13.0
PyJWT==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.6.0
six==1.16.0
SQLAlchemy==1.4.44
stack-data==0.6.2
tomli==2.0.1
traitlets==5.6.0
urllib3==1.26.13
wcwidth==0.2.5
Werkzeug==2.2.2
WTForms==3.0.1
